Controller Base: Fix double initialization in diamond inheritance
This prevents the cartesian_compliance_controller from
initializing the base class twice. The same is not easily possible
with the starting() method, because starting() and stopping() will
get called many times. This could be implemented later, if it becomes
important to not call double starting().
For now, this seems fine.