Joint Cartesian Controller: Further implement functionality
This is a first working version.

However, one of the main purpose of this controller was to use
the rqt joint trajectory controller plugin, which does not work.
That plugin needs an actual JointTrajectoryController, which claims
hardware resources. An alternative approach would be to use an internal
controller manager inside this controller and load a
JointTrajectoryController to transform the joint commands in the
read() and write() functions.