Controller Base: Change time integration to be smooth
This is a design change. The previous approach was stable, but less
smooth. The new time integration leads to smoother motion and is physically
more plausible.
As a consequence, the velocities really reflect the current
simulation state and can be used for damping etc.
This was also discussed in commit e1dceefb41b34f08fb28537631d844a117005dd9