Controller Test: Fix end_effector link for compliance controller testing
Important things to note:
The end_effector_link really must be the last reference link in the
configuration, since the internal robot chain is build up using that
link. If e.g. the end_effector is 'tool0' and the 'sensor_link' is
mounted in the urdf to the 'tool0' link, then this won't work.
It this case, the end_effector must also be set to 'sensor_link'.
The compliance_ref_link can be set to something else, as long as the link
is part of the chain between the base_link and the end_effector_link.
Also remember to adjust the end_effector_link in the
joint_to_cartesian_controller respectively.