Motion Controller: Fix cross correlation bug for translation and rotation
The frame difference really has to be split up in translation and
rotation individually. If this is done in one compact expression, then the
rotation offsets also have effects on the translation offsets.
This cross correlation was sometimes experienced, but it really
hit the eye for the compliance controller. Apparently, the motion
controller limited this with its controller gains, that's why it wasn't
so obvious.