Force and Compliance Controller: Add first attempt for gravity compensation
This enhancement shall compensate the effects of gravity, so that
turning a heavy tool that is mounted on the sensor does not cause
disturbing forces.
The method is as follows:
 - Get notified when the sensor is tared and keep in mind,
   what this taring compensated
 - Apply a virtual force that compensates the current effect of gravity
 - Remove the terms, that were added during the taring process

Todo: Read mass and center of gravity from parameter server.
In future versions, maybe think about automatically measuring the tool.
For now, this should be sufficient.