HACK: set door_0 as fixed frame in tf_lookup
In euroc, door_0 is published much slower than the robot information.
Thus, any tf-lookup including a path over door_0 is cannot be performed
with a higher frequency than the transformation of door_0 is published.

A better solution would probably be to use tf_static to publish the
door_frame as long as it is fixed or to increase the door frame
publishing.