Port `PDController` and `SpacialPDController`
Both change from using internal dynamic reconfigure to the ROS2
parameter mechanism.
We pass a `Node` handle for declaring and retrieving parameters locally.
This handle will later be available from the ROS-controller that
instantiates this code.

From briefly looking at the implementation of getting parameters, this
seems not to be realtime-safe.  We might need to change that later on to
something more suitable.