Allow users to choose time integration method
This only applies to the forward dynamics-based solver.
It's an old feature that I used in the beginning of the
*cartesian_controllers*.
Using a dynamical system and conserving joint velocities during control
is more noise robust due to the double time integration, acting as
double low-pass filter.

It especially makes sense for force and compliance control.

However, it requires additional null space damping and derivative gains
for all Cartesian dimensions to suppress oscillations in free space.