Split synchronization and kinematics updates
We need three functionalities:
(1) Updating the internal end-effector pose/twist with forward
kinematics, (2) synchronizing the internal model with the real robot,
and (3) setting meaningful start states during starting().

The first needs to be done on each simulation step, whereas the second
is suitable after *iterations* passed. The third is required when
starting the controller.