Further limit max. tolerated motion offsets
This is an addition to commit 24f1a9c.
We are now even more stable for all targets outside the robot's reach.
I tested this with every IK solver and it behaves as expected.

Since this applies to the control cycle of the virtual model, we still can
configure an exact IK solver with a higher number of *iterations*.