Revert "Limit max. tolerated motion offsets per control cycle"
This reverts commit 24f1a9cc1046c3ac4254d6ccdf1f91864ee37d39.

As it turns out, limiting the max. tolerated motion offset also limits
the maximal stiffness that the `cartesian_compliance_controller` can
build up. The restoring compliance force is computed with the motion
error and the stiffness. If the motion error is limited, so is the
stiffness, rendering the compliance behavior useless.

We are now back to the previous behavior.