Port the *cartesian_motion_controller* to ROS2
This is analog to the *cartesian_force_controller*.
The distinction between position and velocity actuation via templates is
no longer necessary. The base class takes care of sending control
commands to all supported interfaces.

We also drop the explicit mentioning of the `target_frame_topic`.
That's conveniently solved with remappings.