Add manual test script for joint controller gains
We now have a minimal gui to switch between two hard-coded joint
configurations via a joint trajectory controller.
This is helpful to fine-tune the joint controller gains.

Also lower the robot's link inertias. The previous value of `100` was
unrealistically high. A better next step would be to compute those
values for cylindrical dummies.