Provide safe default command buffers on starting
That's important because the cartesian_controllers' `update()` will only be called
on the *second* cycle of the respective controller manager.
On the *first* cycle, the controllers will only be started and the
underlying robot drivers might write the still untouched command buffers
to the actuators. Depending on the robot driver's implementation
(especially on its command buffer initialization), this could cause
safety critical jumps.

This assures that we provide safe command buffers in the
cartesian_controllers.