Fix callback for link mass in the virtual model
The dynamic parameter `link_mass` is now used as expected in the virtual
model. KDL'joint inertia solver keeps a const reference to our robot
chain and allows us to update it on each call to the forward dynamics
routine.

Also use an increased link masses by default.
That's a stable behavior in singular configurations.