Implement ROS2 publishers for for pose and twist
The active assembly object now reports its current pose and twist with
respect to the world coordinate frame (= the assembly goal frame).

We use a naming convention for the active object (= the body to steer
manually) and exit with an error if that is not found among the bodies
in the xml.